fix: close kanban review regressions
