+++ scripts/ratchet-baseline.json
@@ -241,7 +241,7 @@
     "files": {}
   },
   "lint-suppression-directives": {
-    "count": 743,
+    "count": 742,
     "files": {
       "apps/api/scripts/ai-provider-canary-config.ts": 1,
       "apps/api/scripts/ai-provider-canary.ts": 4,
@@ -514,7 +514,6 @@
       "apps/web/src/routes/_protected.knowledge/-components/template-studio.tsx": 1,
       "apps/web/src/routes/_protected.knowledge/-components/template-wizard.tsx": 1,
       "apps/web/src/routes/_protected.knowledge/clauses.tsx": 1,
-      "apps/web/src/routes/_protected.settings/-components/account/two-factor-card.tsx": 1,
       "apps/web/src/routes/_protected.tsx": 1,
       "apps/web/src/routes/_protected.workspaces/$workspaceId/$viewId.document.tsx": 1,
       "apps/web/src/routes/_protected.workspaces/$workspaceId/$viewId.route.tsx": 1,
